+++ src/app/profile/created-builds/page.tsx
@@ -5,7 +5,6 @@ import BuildCard from '@/features/build/components/BuildCard'
 import usePagination from '@/features/pagination/usePagination'
 import { useEffect, useState } from 'react'
 import { CreatedBuildsFilter, getCreatedBuilds } from '../actions'
-import BuildListFilters from '@/features/build/components/BuildListFilters'
 import CopyBuildUrlButton from '../../../features/profile/CopyBuildUrlButton'
 import EditBuildButton from '../../../features/profile/EditBuildButton'
 import DuplicateBuildButton from '../../../features/profile/DuplicateBuildButton'
@@ -15,14 +14,36 @@ import Tabs from '../../../features/profile/Tabs'
 import ProfileHeader from '../../../features/profile/ProfileHeader'
 import { useSession } from 'next-auth/react'
 import AuthWrapper from '@/features/auth/components/AuthWrapper'
+import useBuildListFilters from '@/features/filters/hooks/useBuildListFilters'
+import BuildListFilters from '@/features/filters/components/BuildListFilters'
+import CommunityBuildFilters from '@/features/filters/components/CommunityBuildFilters'
+import { CommunityBuildFilterProps } from '@/features/filters/types'
+import { DEFAULT_COMMUNITY_BUILD_FILTERS } from '@/features/filters/constants'
 
 export default function Page() {
   const { data: sessionData } = useSession()
   const [builds, setBuilds] = useState<DBBuild[]>([])
   const [totalBuildCount, setTotalBuildCount] = useState<number>(0)
-  const [filter, setFilter] = useState<CreatedBuildsFilter>('date created')
   const [isLoading, setIsLoading] = useState(false)
+
+  const [communityBuildFilters, setCommunityBuildFilters] =
+    useState<CommunityBuildFilterProps>(DEFAULT_COMMUNITY_BUILD_FILTERS)
+  function handleChangeCommunityBuildFilters(
+    filters: CommunityBuildFilterProps,
+  ) {
+    setCommunityBuildFilters(filters)
+  }
+
   const itemsPerPage = 16
+
+  const {
+    orderBy,
+    orderByOptions,
+    timeRange,
+    timeRangeOptions,
+    handleOrderByChange,
+    handleTimeRangeChange,
+  } = useBuildListFilters()
 
   const {
     currentPage,
@@ -42,22 +63,18 @@ export default function Page() {
     const getItemsAsync = async () => {
       setIsLoading(true)
       const response = await getCreatedBuilds({
+        communityBuildFilters,
         itemsPerPage,
+        orderBy,
         pageNumber: currentPage,
-        filter,
+        timeRange,
       })
       setBuilds(response.items)
       setTotalBuildCount(response.totalItemCount)
       setIsLoading(false)
     }
     getItemsAsync()
-  }, [currentPage, itemsPerPage, filter])
-
-  const filterOptions: CreatedBuildsFilter[] = ['date created', 'upvotes']
-
-  function handleFilterChange(filter: string) {
-    setFilter(filter as CreatedBuildsFilter)
-  }
+  }, [communityBuildFilters, currentPage, itemsPerPage, orderBy, timeRange])
 
   function handleDeleteBuild(buildId: string) {
     setBuilds((prevBuilds) =>
@@ -77,6 +94,9 @@ export default function Page() {
       <div className="mb-8 flex w-full flex-col items-center">
         <Tabs />
       </div>
+      <div className="mb-8 flex w-full max-w-2xl items-center justify-center">
+        <CommunityBuildFilters onUpdate={handleChangeCommunityBuildFilters} />
+      </div>
       <BuildList
         label="Builds you've created"
         currentPage={currentPage}
@@ -90,13 +110,14 @@ export default function Page() {
         onNextPage={handleNextPageClick}
         onSpecificPage={handleSpecificPageClick}
         headerActions={
-          <div className="min-w-[150px] max-w-[250px]">
-            <BuildListFilters
-              filter={filter}
-              onFilterChange={handleFilterChange}
-              options={filterOptions}
-            />
-          </div>
+          <BuildListFilters
+            orderBy={orderBy}
+            orderByOptions={orderByOptions}
+            onOrderByChange={handleOrderByChange}
+            timeRange={timeRange}
+            timeRangeOptions={timeRangeOptions}
+            onTimeRangeChange={handleTimeRangeChange}
+          />
         }
       >
         {builds.map((build) => (