+++ src/app/profile/favorited-builds/page.tsx
@@ -4,7 +4,6 @@ import CopyBuildUrlButton from '../../../features/profile/CopyBuildUrlButton'
 import usePagination from '@/features/pagination/usePagination'
 import { useEffect, useState } from 'react'
 import BuildCard from '@/features/build/components/BuildCard'
-import BuildListFilters from '@/features/build/components/BuildListFilters'
 import BuildList from '@/features/build/components/BuildList'
 import DuplicateBuildButton from '../../../features/profile/DuplicateBuildButton'
 import { DBBuild } from '@/features/build/types'
@@ -12,15 +11,37 @@ import Tabs from '../../../features/profile/Tabs'
 import ProfileHeader from '../../../features/profile/ProfileHeader'
 import { useSession } from 'next-auth/react'
 import AuthWrapper from '@/features/auth/components/AuthWrapper'
-import { SortFilter, getFavoritedBuilds } from './actions'
+import { getFavoritedBuilds } from './actions'
+import CommunityBuildFilters from '@/features/filters/components/CommunityBuildFilters'
+import { CommunityBuildFilterProps } from '@/features/filters/types'
+import { DEFAULT_COMMUNITY_BUILD_FILTERS } from '@/features/filters/constants'
+import useBuildListFilters from '@/features/filters/hooks/useBuildListFilters'
+import BuildListFilters from '@/features/filters/components/BuildListFilters'
 
 export default function Page() {
   const { data: sessionData } = useSession()
   const [builds, setBuilds] = useState<DBBuild[]>([])
   const [totalBuildCount, setTotalBuildCount] = useState<number>(0)
   const [isLoading, setIsLoading] = useState(false)
-  const [sortFilter, setSortFilter] = useState<SortFilter>('date favorited')
+
+  const [communityBuildFilters, setCommunityBuildFilters] =
+    useState<CommunityBuildFilterProps>(DEFAULT_COMMUNITY_BUILD_FILTERS)
+  function handleChangeCommunityBuildFilters(
+    filters: CommunityBuildFilterProps,
+  ) {
+    setCommunityBuildFilters(filters)
+  }
+
   const itemsPerPage = 16
+
+  const {
+    orderBy,
+    orderByOptions,
+    timeRange,
+    timeRangeOptions,
+    handleOrderByChange,
+    handleTimeRangeChange,
+  } = useBuildListFilters()
 
   const {
     currentPage,
@@ -40,22 +61,18 @@ export default function Page() {
     const getItemsAsync = async () => {
       setIsLoading(true)
       const response = await getFavoritedBuilds({
+        communityBuildFilters,
         itemsPerPage,
+        orderBy,
         pageNumber: currentPage,
-        sortFilter,
+        timeRange,
       })
       setBuilds(response.items)
       setTotalBuildCount(response.totalItemCount)
       setIsLoading(false)
     }
     getItemsAsync()
-  }, [currentPage, itemsPerPage, sortFilter])
-
-  const filterOptions: SortFilter[] = ['date favorited', 'upvotes']
-
-  function handleFilterChange(filter: string) {
-    setSortFilter(filter as SortFilter)
-  }
+  }, [communityBuildFilters, currentPage, itemsPerPage, orderBy, timeRange])
 
   return (
     <AuthWrapper>
@@ -68,6 +85,9 @@ export default function Page() {
       )}
       <div className="mb-8 flex w-full flex-col items-center">
         <Tabs />
+      </div>
+      <div className="mb-8 flex w-full max-w-2xl items-center justify-center">
+        <CommunityBuildFilters onUpdate={handleChangeCommunityBuildFilters} />
       </div>
       <BuildList
         label="Builds you've favorited"
@@ -82,13 +102,14 @@ export default function Page() {
         onNextPage={handleNextPageClick}
         onSpecificPage={handleSpecificPageClick}
         headerActions={
-          <div className="min-w-[150px] max-w-[250px]">
-            <BuildListFilters
-              filter={sortFilter}
-              onFilterChange={handleFilterChange}
-              options={filterOptions}
-            />
-          </div>
+          <BuildListFilters
+            orderBy={orderBy}
+            orderByOptions={orderByOptions}
+            onOrderByChange={handleOrderByChange}
+            timeRange={timeRange}
+            timeRangeOptions={timeRangeOptions}
+            onTimeRangeChange={handleTimeRangeChange}
+          />
         }
       >
         {builds.map((build) => (